backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-jose[cryptography]
passlib
bcrypt==4.0.1
cryptography
python-dotenv
qrcode
Pillow